+++ junit-common/src/test/java/eu/coldrye/junit/util/ClassPathResourceTest.java
@@ -1,7 +1,0 @@
-// TBD:LICENSE
-
-package eu.coldrye.junit.util;
-
-public class ClassPathResourceTest {
-
-}